refactor: reorganize form folders
