+++ src/pages/home/form/HookForm.tsx
@@ -1,12 +1,12 @@
-import '../plain_form/plain_form.css';
+import './form.css';
 import { useNavigate } from 'react-router';
 import { SubmitHandler, useForm } from 'react-hook-form';
 import { useSelector, useDispatch } from 'react-redux';
-import { RootState } from '../../state/store';
+import { RootState } from '../../../state/store';
 import { z } from 'zod';
 import { zodResolver } from '@hookform/resolvers/zod';
-import { convertFileToBase64 } from '../../utils/convertFile';
-import { addFormData } from '../../state/features/form/formDataSlice';
+import { convertFileToBase64 } from '../../../utils/convertFile';
+import { addFormData } from '../../../state/features/form/formDataSlice';
 
 const fileSizeLimit = 5 * 1024 * 1024;
 const schema = z
